extract weatherdays component and clean up unused code and import
